Extract shared test fixtures (MockResponse, client, local_client) to conftest.py
Centralizes reusable test infrastructure in a shared conftest module,
removing duplication from test_client.py and adding type hints.
